backend/requirements.txt
setuptools>=65.0.0
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
PyMySQL==1.1.0
openpyxl==3.1.2
pandas>=1.5.0
python-dotenv==1.0.0
